fix(StoreQueue): setting addrvalid on misalign revoke (#4650)
Previously, we set addrvalid to be set only at s1.

For misaligned stores, if it will go into misalignbuffer at s1, it may
generate a revoke at s2 due to an exception.
If no revoke is generated, the addrvalid should not be set because the
addrvalid will be set by the subsequent split request.
And if a revoke is generated, then an exception is generated and the
addrvalid should be set at s2 to make this uop available for enq.