fix(LoadQueueUncache): return origin data in LoadQueueUncache. (#4703)
**Bug Trigger:** When the data responsed by `Uncache` is 0x0013 xxxx
xxxx xxxx, the selected data handled by `LoadQueueUncache` is `0x0000
0000 0000 0013` in which case that `raddr(2, 0)` is 0b110 and
`LSUOpType` is `lhu`. And this data is what we want. However it is then
processed by the LoadUnit again with the same process and a new data
`0x0000 0000 0000 0013` is obtained, which is error.

So here needs to return origin data which will be selected in LoadUnit.