fix: safe protection before entering of WFI (backend/ifu/mmu/lsu) (#4691)
fix: add safe protection according to the original semantics of WFI

* This fix tries to solve the unsafe scenario where the Core is in the
'WFI' state, but there
are still unfinished L2 transactions coming from IFU/MMU/LSU, which
violates the original
semantics of WFI. So, this protection was designed to guarantee that
there are no pending
      requests to L2 before the Core enters the 'WFI' state.

* The BE (Backend) actively initiates the reqWfi=1 signal to notify the
IFU/MMU/LSU that the
      Core wants to enter WFI.

* For IFU/MMU/LSU, when reqWfi=1, they stop sending all requests to L2.
After all pending
      requests are completed, they send a safe=1 signal back to BE.

* If an interrupt occurs while BE is waiting for the safe signal, it
directly pulls reqWfi=0
to notify IFU/MMU/LSU that they do not need to enter the safe state, and
BE automatically ends.
At the same time, if IFU/MMU/LSU discover reqWfi=0 while waiting for
pending requests, they
remove the restriction on sending requests to L2 and pull the safe
signal low.

* The writeBack request from LSU to L2 is NOT included in this safe
handshake protocol, because
in the multi-core non-low-power WFI scenario, there may be a snoop from
another Core to the
'WFI' Core. In this case, this writeback channel should be ACTIVE to
respond to this snoop.
However, in the low-power 'WFI' state, the 'WFI' Core will exit
coherency before power down,
      so there will be no this kind of snoop.
   
* To avoid timing iteration, each wfiSafe from ifu/mmu/lsu is registered
before send to BE
   
* Currently, this feature is fixed and verified by all regression tests
on PLDM

---------

Co-authored-by: chengguanghui <wissycgh@gmail.com>
Co-authored-by: ngc7331 <ngc7331@outlook.com>
Co-authored-by: sfencevma <15155930562@163.com>
Co-authored-by: good-circle <fenghaoyuan19@mails.ucas.ac.cn>