fix(VMergeBuffer): debug mode is also treated as an exception (#4707)
Triggers include debug mode and breakpoint exceptions.

The breakpoint exception will be in the exceptionVec, so no additional
processing is required.
Whereas debug mode is encoded in the signal trigger, when the trigger is
set to debug mode, the processor should respond as if it were an
exception, except that it will go into debug mode instead of an
exception.