fix(StoreQueue): cbozero flag is set using `wline` flag (#4759)
We no longer use deqptr to index fuoptype for cbozero determination, as
this can lead to errors when a misaligned store is adjacent to cbozero:
An misaligned store is split into two sbuffer requests. 
Assuming deqptr(0) is an unaligned store and deqptr(1) is cbozero, after
the non-aligned split, both sbuffer(0) and sbuffer(1) are requests for
the misaligned store.
Using deqptr for indexing would incorrectly identify sbuffer(1) as
cbozero.

Therefore, we use the wline flag for determination. Currently, the wline
flag is only set by cboinstructions, and cboinstructions other than
cbozerodo not enter the sbuffer, so this only applies to cbozero.