fix(MetaArray): enable write bypass for read (#4791)
Enable bypass to allow reads to obtain write data when reads and writes
occur simultaneously.

---

load s0 reads meta.
load s1 obtains the read meta result.
load s2 requests entry into rarqueue.

If a write meta probe is generated simultaneously with load s0, without
enabling bypass, this read meta will obtain old data.
This will result in a dcache hit.
However, the rarqueue release occurs on the next cycle after the probe
write meta, which is when load s1 is generated. At this point, the load
has not yet entered the rarqueue, so it is not marked as released in the
rarqueue.