fix(LSU): fix misalignbuffer response acceptance process (#4840)
Since our misaligned wakeup is generated by three beats in place in the
load unit, once the misaligned wakeup request enters the load unit, it
cannot be flushed.
Therefore, it is possible that this instruction has been redirected, but
a response has been sent back from the load unit, causing the misalign
buffer to accept a response that it should not accept.

fix:
For simplicity, we make a judgement in the misalign buffer and do not
accept responses from non-current instructions.